backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
google-genai>=1.74.0
pymupdf==1.24.10
python-dotenv==1.0.1
pydantic==2.7.4
pillow==10.4.0
